Remove list-based transpose function.
I tinkered with replacing the Vector based transpose function with a
list-based one. At first this seemed an opportunity to simplify some
proofs, but it appears that too many other proofs get significantly
more complex from the propagation of default values and dimensional
properties. I think the current solution is a better compromise.