Add determine_context_word.
I tried to outline a different approach for this rather hairy proof.
The cast function is used to convert a list into a matrix such that
transpose can be used to compute the context word. Now the rest of
the proof consists of showing that transpose can invert itself, and
showing that encode and decode cancel out.