Replace Similar with Automorphism.
Although the automorphism method works, I realized that it is not really
worth the trouble to work with canonical states. Instead I want to try
using Eqdep_dec to define the powerset construction using states that
are canonical (sorted and deduplicated) by default. If this works all
states will be finite types and we still won't need any axioms.