Progress regular_ex.
This theorem is set to become the most involved one of the entire
project so far. The repeated conversion between lists and vectors
is annoying, but previous attempts at removing vectors ended up in
an increase of proof complexity. What is both very useful and very
horrible about vectors is that their length is part of the term.
This becomes a nightmare when rewriting over list-to-vector
conversion, because types are often inconvertible. But each time
it turned out that there is a way out by carefully selecting
intermediate targets.