Prove dfs_complete.
This is quite a monstrous proof. At this point it starts making more
sense to rely on something like math-comp to deal with graphs. Their
proofs seem more compact, but I do not know how they work. The main
difficulty of my DFS algorithm is that the visited nodes are
propagated, such that each node is really only visited once.

I kept the soundess proof short at the cost of being a bit weaker;
it does not show that the resulting path avoids already visited nodes,
only that the returned path is a valid one.