Add Increasing_In_powerset.
Of course this commit adds a number of other theorems as well, but
this is the main result. The way hypotheses of normalization are
satisfied is ugly, in particular because they need to be supplied
every time a theorem about normalization is used. But I do not
know a truly modular way to avoid this.