Prove various orderings correct.
This completes an upgrade of the previous algorithm from a naive
powerset list normalization (which required the whole state space to
be listed explicitly) to a more efficient one that uses merge-sort
to sort lists of states using a linear order.

The resulting algorithm is indeed faster. A formula that previously took
around 20 seconds to solve is now solved in an instant. But this success
is limited to only a few cases; slightly larger formulas are still
infeasible.

Only a few obvious optimizations are left that could improve performance
inside Coq, but I am not sure if they are worthwhile. Now that we have a
comparison function for states, it is possible to store the visited
states of the depth-first search in an AVL-tree. It is also possible that
separating the algorithm from its specification could improve runtime
performance, but this would also make the proof less readable (although
the algorithm would become more explicit). Extracting OCaml code and
testing its performance would be interesting as well, although I think
it is safe to say the algorithm is useless for practical applications.